+++ utils/ip/ip.go
@@ -4,7 +4,7 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-package helpers
+package ip
 
 import "net"
